Search Lichess user's rating history
Code for running the functions defined in scraping_lichess file. Accesses Lichess API, allows the user to search for information on the current top ten rated players for each game type or to search for the rating history of a specific user. Specific user's rating history is added to a database and queried with sqlite3 to allow the results to be filtered by date or by rating for classic, correspondence or puzzle game types.